+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Removed
+
+- `checkboxes` prop from `EXPERIMENTAL_Table`.
+- `EXPERIMENTAL_TableTree` component.
+
 ## [9.112.20] - 2020-03-17
 
 ### Fixed